fix: resolve cyclic dependencies and domain import violations
## Summary
- Fixed 10 cyclic dependencies detected in the application layer
- Fixed 8 domain import violations
- Updated dependency checker to properly exclude TYPE_CHECKING and lazy imports

## Changes

### Dependency Checker Improvements
- Added `bioetl.domain.value_objects` and `bioetl.domain.data` to allowed
  domain imports (legitimate dependencies for pipelines)
- Modified `build_dependency_graph` to exclude TYPE_CHECKING and lazy imports
  from cycle detection since they don't create runtime cycles

### Architectural Fixes
1. **Moved hooks_impl.py to factories/**
   - Breaks the pipelines -> factories -> pipelines cycle
   - Old location re-exports for backward compatibility

2. **Moved executor.py into pipelines/**
   - Breaks the pipelines -> executor -> pipelines cycle
   - executor logically belongs with pipeline execution code
   - Old location re-exports for backward compatibility

### Result
All dependency checks now pass:
- No cyclic dependencies
- No infrastructure import violations
- No forbidden imports
- All domain imports from allowed list
